Remapped Driver controls, and began work on implementing an Accelerometer.
At this moment, the DriveXInchesCommand is not working and we are going to begin to implement an Accelerometer in order to ensure that the position of the robot is accurate. The Claw subsystem has had the bottom level decreased in order to be flush against the bottom of the robot.